Added search feature with search results page hitting scryfall api

+++ FortyLife.Data/Scryfall/Card.cs
@@ -37,7 +37,7 @@ namespace FortyLife.Data.Scryfall
 
         public string ManaCost { get; set; }
 
-        public int Cmc { get; set; }
+        public double Cmc { get; set; }
 
         public string TypeLine { get; set; }
